+++ src/builders.ts
@@ -23,6 +23,7 @@ import * as normalizePackageData from 'normalize-package-data'
 import { type PackageURL } from 'packageurl-js'
 import * as path from 'path'
 
+import { loadJsonFile } from './_helpers'
 import { makeNpmRunner, type runFunc } from './npmRunner'
 import { PropertyNames, PropertyValueBool } from './properties'
 import { versionCompare } from './versionCompare'
@@ -350,8 +351,7 @@ export class BomBuilder {
     const packageJsonPath = path.join(data.path, 'package.json')
     try {
       return Object.assign(
-        /* eslint-disable-next-line @typescript-eslint/no-var-requires */
-        require(packageJsonPath),
+        loadJsonFile(packageJsonPath),
         data
       )
     } catch {
@@ -590,7 +590,7 @@ export class BomBuilder {
 
     for (const packageJsonPath of packageJsonPaths) {
       /* eslint-disable-next-line @typescript-eslint/no-var-requires */
-      const packageData = require(packageJsonPath)
+      const packageData = loadJsonFile(packageJsonPath)
       normalizePackageData(packageData /* add debug for warnings? */)
       const tool = this.toolBuilder.makeTool(packageData)
       if (tool !== undefined) {