+++ src/builders.ts
@@ -166,7 +166,6 @@ export class BomBuilder {
       }
     }
 
-    // TODO use instead ? : https://www.npmjs.com/package/debug ?
     this.console.info('INFO  | gather dependency tree ...')
     this.console.debug('DEBUG | npm-ls: run npm with %j in %j', args, projectDir)
     let npmLsReturns: Buffer
@@ -208,7 +207,6 @@ export class BomBuilder {
   }
 
   buildFromNpmLs (data: any, npmVersion?: string): Models.Bom {
-    // TODO use instead ? : https://www.npmjs.com/package/debug ?
     this.console.info('INFO  | build BOM ...')
 
     // region all components & dependencies
@@ -310,11 +308,13 @@ export class BomBuilder {
     for (const [depName, depData] of Object.entries(data.dependencies ?? {}) as any) {
       if (depData === null || typeof depData !== 'object') {
         // cannot build
+        this.console.debug('DEBUG | skip malformed component %j in %j', depName, depData)
         continue // for-loop
       }
       if (typeof depData.path !== 'string') {
         // might be an optional dependency that was not installed
         // skip, as it was not installed anyway
+        this.console.debug('DEBUG | skip missing component %j in %j', depName, depData.path)
         continue // for-loop
       }
 
@@ -323,14 +323,17 @@ export class BomBuilder {
         const _dep = this.makeComponent(depData)
         if (_dep === false) {
           // shall be skipped
+          this.console.debug('DEBUG | skip impossible component %j in %j', depName, depData.path)
           continue // for-loop
         }
         dep = _dep ??
           new DummyComponent(Enums.ComponentType.Library, `InterferedDependency.${depName as string}`)
         if (dep instanceof DummyComponent) {
-          this.console.warn('WARN  | InterferedDependency $j', dep.name)
+          this.console.warn('WARN  | InterferedDependency %j in %j', depName, depData.path)
+        } else {
+          this.console.debug('DEBUG | built component %j in %j: %j', depName, depData.path, dep)
         }
-
+        this.console.info('INFO  | add component for %j in %j', depName, depData.path)
         allComponents.set(depData.path, dep)
       }
       directDepRefs.add(dep.bomRef)