+++ HISTORY.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 
 ## unreleased
 
+* Changed
+  * Log output is less verbose, can be re-enabled via CLI switch `--verbose` ([#158] via [#1131])  
+    Warnings and errors are still displayed as before.  
+    This is considered a non-breaking change, since only informational logs and debug information is affected.
+* Added
+  * CLI switch `-v, --verbose` to increate output verbosity ([#158] via [#1131])  
+    May be used multiple times, like `-vvv`.
+
+[#158]: https://github.com/CycloneDX/cyclonedx-node-npm/issues/158
+[#1131]: https://github.com/CycloneDX/cyclonedx-node-npm/pull/1131
+
 ## 1.14.3 - 2023-12-01
 
 * Fixed