+++ HISTORY.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## unreleased
 
+* Changed
+  * Hardened JSON imports (via [#1132]) 
+
+[#1132]: https://github.com/CycloneDX/cyclonedx-node-npm/pull/1132
+
 ## 1.14.3 - 2023-12-01
 
 * Fixed