+++ HISTORY.md
@@ -5,8 +5,18 @@ All notable changes to this project will be documented in this file.
 ## unreleased
 
 * Changed
-  * Hardened JSON imports (via [#1132]) 
+  * Log output is less verbose, can be re-enabled via CLI switch `--verbose` ([#158] via [#1131])  
+    Warnings and errors are still displayed as before.  
+    This is considered a non-breaking change, since only informational logs and debug information is affected.
+  * Hardened JSON imports (via [#1132])
+* Added
+  * CLI switch `-v, --verbose` to increase output verbosity ([#158] via [#1131])  
+    May be used multiple times, like `-vvv`.
+  * More logs on info-level (via [#1131])
+  * More logs on debug-level (via [#1131])
 
+[#158]: https://github.com/CycloneDX/cyclonedx-node-npm/issues/158
+[#1131]: https://github.com/CycloneDX/cyclonedx-node-npm/pull/1131
 [#1132]: https://github.com/CycloneDX/cyclonedx-node-npm/pull/1132
 
 ## 1.14.3 - 2023-12-01